+++ src/shellyTypes.ts
@@ -27,6 +27,40 @@ export type ShellyData = Record<string, ShellyDataType>;
 
 export type ShellyDataType = string | number | boolean | null | undefined | object;
 
+export interface Pm1Meter {
+  id: number;
+  voltage: number;
+  current: number;
+  apower: number;
+  freq: number;
+  aenergy: {
+    total: number;
+    by_minute: number[];
+    minute_ts: number;
+  };
+  ret_aenergy: {
+    total: number;
+    by_minute: number[];
+    minute_ts: number;
+  };
+}
+
+export interface Em1Meter {
+  id: number;
+  act_power: number;
+  aprt_power: number;
+  current: number;
+  freq: number;
+  pf: number;
+  voltage: number;
+}
+
+export interface Em1MeterData {
+  id: number;
+  total_act_energy: number;
+  total_act_ret_energy: number;
+}
+
 // Define the interface for the BTHomeDeviceComponent status object
 export interface BTHomeDeviceComponentStatus {
   id: number;
@@ -91,4 +125,28 @@ export interface BTHomeComponent {
   key: string;
   status: BTHomeDeviceComponentStatus | BTHomeSensorComponentStatus;
   config: BTHomeDeviceComponentStatus | BTHomeSensorComponentConfig;
+}
+
+// Define the interface for BTHomeEvent from the BTHomeDevice on WebSocket
+export interface BTHomeEvent {
+  component: string;
+  id: number;
+  event: string;
+  ts: number;
+  data?: ShellyData;
+}
+
+// Define the interface for WebSocket NotifyStatus from a BTHomeDevice
+export interface BTHomeStatusDevice {
+  id: number; // The BTHome device id: 200, 201 etc.
+  last_updated_ts: number;
+  packet_id: number;
+  rssi: number;
+}
+
+// Define the interface for WebSocket NotifyStatus from a BTHomeSensor
+export interface BTHomeStatusSensor {
+  id: number; // The BTHome device id: 200, 201 etc.
+  last_updated_ts: number;
+  value?: ShellyDataType;
 }